# docs/swagger.yaml
basePath: /
definitions:
  forms.CreatePostBody:
    properties:
      content:
        type: string
      imgUrl:
        type: string
      tags:
        description: JSON 数组
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - content
    - imgUrl
    - tags
    - title
    type: object
  forms.PostItem:
    properties:
      adjustTime:
        description: 格式化后的时间
        type: string
      id:
        type: integer
      imgUrl:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  forms.PostResponse:
    properties:
      adjustTime:
        type: string
      content:
        type: string
      id:
        type: integer
      imgUrl:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  forms.PostsPage:
    properties:
      list:
        items:
          $ref: '#/definitions/forms.PostItem'
        type: array
      total:
        type: integer
    type: object
  forms.SearchPosts:
    properties:
      page:
        minimum: 1
        type: integer
      pageSize:
        maximum: 100
        minimum: 1
        type: integer
      q:
        type: string
    required:
    - page
    - pageSize
    type: object
  utils.ErrorResponse:
    properties:
      code:
        type: integer
      errorMessage:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: xinghe
    url: http://www.swagger.io/support
  description: 基于 Gin 的博客后端 API
  termsOfService: http://swagger.io/terms/
  title: 我的博客 API
  version: "1.0"
paths:
  /blog:
    post:
      consumes:
      - application/json
      description: 根据传入的参数创建一篇新文章
      parameters:
      - description: 文章信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/forms.CreatePostBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forms.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: 创建文章
      tags:
      - blog
  /search:
    post:
      consumes:
      - application/json
      description: 根据传入的参数搜索文章
      parameters:
      - description: 文章关键字
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/forms.SearchPosts'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forms.PostsPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: 搜索文章
      tags:
      - blog
swagger: "2.0"